# Add RigReprojErrorConstantRigCostFunction