--- a/src/colmap/estimators/cost_functions_test.cc
+++ b/src/colmap/estimators/cost_functions_test.cc
@@ -30,6 +30,7 @@
 #include "colmap/estimators/cost_functions.h"
 
 #include "colmap/geometry/pose.h"
+#include "colmap/geometry/rigid3.h"
 #include "colmap/sensor/models.h"
 
 #include <gtest/gtest.h>
@@ -165,6 +166,43 @@ TEST(BundleAdjustment, Rig) {
   const double* parameters[6] = {cam_from_rig_rotation,
                                  cam_from_rig_translation,
                                  rig_from_world_rotation,
+                                 rig_from_world_translation,
+                                 point3D,
+                                 camera_params};
+  EXPECT_TRUE(cost_function->Evaluate(parameters, residuals, nullptr));
+  EXPECT_EQ(residuals[0], 0);
+  EXPECT_EQ(residuals[1], 0);
+
+  point3D[1] = 1;
+  EXPECT_TRUE(cost_function->Evaluate(parameters, residuals, nullptr));
+  EXPECT_EQ(residuals[0], 0);
+  EXPECT_EQ(residuals[1], 1);
+
+  camera_params[0] = 2;
+  EXPECT_TRUE(cost_function->Evaluate(parameters, residuals, nullptr));
+  EXPECT_EQ(residuals[0], 0);
+  EXPECT_EQ(residuals[1], 2);
+
+  point3D[0] = -1;
+  EXPECT_TRUE(cost_function->Evaluate(parameters, residuals, nullptr));
+  EXPECT_EQ(residuals[0], -2);
+  EXPECT_EQ(residuals[1], 2);
+}
+
+TEST(BundleAdjustment, ConstantRig) {
+  Rigid3d cam_from_rig;
+  cam_from_rig.translation << 0, 0, -1;
+  LOG(INFO) << cam_from_rig.rotation << cam_from_rig.translation;
+  std::unique_ptr<ceres::CostFunction> cost_function(
+      RigReprojErrorConstantRigCostFunction<SimplePinholeCameraModel>::Create(
+          cam_from_rig, Eigen::Vector2d::Zero()));
+
+  double rig_from_world_rotation[4] = {0, 0, 0, 1};
+  double rig_from_world_translation[3] = {0, 0, 1};
+  double point3D[3] = {0, 0, 1};
+  double camera_params[3] = {1, 0, 0};
+  double residuals[2];
+  const double* parameters[4] = {rig_from_world_rotation,
                                  rig_from_world_translation,
                                  point3D,
                                  camera_params};