@ahojnnes 2024-02-01
Thanks a lot for the benchmark numbers. I guess the difference is marginal. With the benchmark setup, it's also a little unclear, if the cost function or the constructor leads to the increase, as you only evaluate it once.