@ahojnnes 2024-02-01
I will take another look later but then probably just merge.