@sarlinpe 2024-02-01
Updated the benchmark code to avoid measuring the construction:
```
main branch
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
BM_ReprojErrorCostFunction/Run_mean                       51161 ns        51161 ns           50
BM_ReprojErrorCostFunction/Run_median                     51114 ns        51113 ns           50
BM_ReprojErrorCostFunction/Run_stddev                       170 ns          170 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_mean           39244 ns        39244 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_median         39217 ns        39216 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_stddev           122 ns          122 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_mean        44527 ns        44527 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_median      44493 ns        44493 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_stddev        177 ns          177 ns           50

This PR - no inlining
-----------------------------------------------------------------------------------------------
Benchmark                                                     Time             CPU   Iterations
-----------------------------------------------------------------------------------------------
BM_ReprojErrorCostFunction/Run_mean                       51454 ns        51454 ns           50
BM_ReprojErrorCostFunction/Run_median                     51417 ns        51417 ns           50
BM_ReprojErrorCostFunction/Run_stddev                       185 ns          185 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_mean           40587 ns        40587 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_median         40565 ns        40565 ns           50
BM_ReprojErrorConstantPoseCostFunction/Run_stddev           128 ns          128 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_mean        45538 ns        45537 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_median      45486 ns        45485 ns           50
BM_ReprojErrorConstantPoint3DCostFunction/Run_stddev        159 ns          159 ns           50
```
The numbers are mostly unchanged.